backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
psycopg[binary]
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
daphne==4.0.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
django-filter==23.2
django-filter==23.2
